move entity and claim to datatypes.py
